moved change password to profile screen
